requirements.txt
PyPDF2==3.0.1
qdrant-client==1.1.5
numpy>=1.21
sentence-transformers==2.2.2
flask
boto3
tqdm
anthropic
threading